problems with long

+++ app/src/main/java/com/stom/videoseekerproject/Video.java
@@ -4,9 +4,9 @@ package com.stom.videoseekerproject;
  * Created by stom.k on 12.05.2015.
  */
 public class Video {
-    public int duration;
+    public long duration;
 
-    public Video(int duration) {
+    public Video(long duration) {
         this.duration = duration;
     }
 }
